+++ app/api/externalIntegrations.v2/automaticTranslation/AutomaticTranslationFactory.ts
@@ -15,7 +15,7 @@ import { ATExternalAPI } from './infrastructure/ATExternalAPI';
 import { MongoATConfigDataSource } from './infrastructure/MongoATConfigDataSource';
 import { Validator } from './infrastructure/Validator';
 import { ATTaskMessage, RequestEntityTranslation } from './RequestEntityTranslation';
-import { SaveEntityTranslationPending } from './SaveEntityTranslationPending';
+import { SaveEntityTranslationPending } from './SaveEntityTranslationsPending';
 import { SaveEntityTranslations } from './SaveEntityTranslations';
 import { ATConfigService } from './services/GetAutomaticTranslationConfig';
 import { SemanticConfig, semanticConfigSchema } from './types/SemanticConfig';
